requirements.txt
# Flask and extensions
flask==3.0.0
flask-login==0.6.3
flask-caching==2.1.0
flask-sqlalchemy==3.1.1
flask-mail==0.9.1
flask-limiter==3.5.0

# Task scheduling and async
apscheduler==3.10.4
httpx==0.27.2

# API and networking
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0

# Database and storage
Werkzeug==3.0.1
SQLAlchemy==2.0.23
alembic==1.13.0

# Security and encryption
bcrypt==4.1.2
cryptography==41.0.7
PyJWT==2.8.0

# Data processing
numpy>=1.24.0
pandas>=2.1.3

# Monitoring and logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Development and testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Production
gunicorn==21.2.0
gevent==23.9.1
